+++ model/elasticsearch/edition.py
@@ -1,6 +1,6 @@
-from elasticsearch_dsl import Text, Keyword, Date, Integer, Nested
+from elasticsearch_dsl import Text, Keyword, Date, Integer, Nested, Object
 
-from .base import BaseInner, plain_ascii
+from .base import BaseInner, PerLanguageField
 from .language import Language
 from .identifier import Identifier
 from .agent import Agent
@@ -8,8 +8,8 @@ from .rights import Rights
 
 
 class Edition(BaseInner):
-    title = Text(analyzer=plain_ascii, fields={'keyword': Keyword()})
-    sub_title = Text(analyzer=plain_ascii, fields={'keyword': Keyword()})
+    title = Object(PerLanguageField)
+    sub_title = Object(PerLanguageField)
     alt_titles = Text(fields={'keyword': Keyword()})
     publication_place = Text(fields={'keyword': Keyword()})
     publication_date = Date(format='date_optional_time')
@@ -37,7 +37,7 @@ class Edition(BaseInner):
 
     def __dir__(self):
         return ['agents', 'identifiers', 'rights', 'languages', 'formats']
-    
+
     def cleanRels(self):
         for rel in dir(self):
             if isinstance(getattr(self, rel), set):